Title

Title

Usage

c_do_adjustRtime_peakGroups(
  peaks,
  peakIndex,
  rtime,
  minFraction = 0.9,
  extraPeaks = 1,
  smooth = c("loess", "linear"),
  span = 0.2,
  family = c("gaussian", "symmetric"),
  peakGroupsMatrix = matrix(ncol = 0, nrow = 0),
  subset = integer(),
  subsetAdjust = c("average", "previous")
)

Arguments

peaks

The identified peaks and their positions

peakIndex

index of peak

rtime

retention times of samples

minFraction

the minimum number of samples a feature has to exist in to qualify as an "aligning feature"

extraPeaks

How many extra peaks are allowed in the area of alignment (can be high for large datasets with high drift)

smooth

smoothing of alignment

span

how many features should the non-linear alignment model look at simultaniously? close to 0 models highly non-linear drift and close to 1 models linear drift

family

kind of model

peakGroupsMatrix

the groups (used for minFraction)

subset

The code is rewritten such that the subset option is no longer available. All samples will be aligned.

subsetAdjust

see "subset"

Value

adjusted rtimes